Work on comments on PR : iter 1. Left to do : Account for tables and graphs.
