+++ tsconfig.json
@@ -8,6 +8,7 @@
 		],
 		"sourceMap": true,
 		"rootDir": ".",
+		"noUnusedLocals": true,
 		"experimentalDecorators": true
 	},
 	"exclude": [
